Visitors can search for images
I had to refactor the search.rb class heavily to extend it. 
There's a lot of heavy lifting that can be handled by the RBing gem.
It might be a good idea to refactor search_spec.rb and the class if/when we refactor to FakeWeb.

Also, Net::HTTP is slow. If we experience performance problems we should consider alternative ruby HTTP libs.