Lowercase SAYT suggestions before validating
- Without doing this, ActiveRecord does a case SENSITIVE comparison when checking for uniqueness during the validation. Documentation exists that claims both TRUE and FALSE for this:
http://ar.rubyonrails.org/classes/ActiveRecord/Validations/ClassMethods.html#M000086
:case_sensitive - Looks for an exact match. Ignored by non-text columns (false by default).
http://www.gotapi.com/rubyrails
:case_sensitive - Looks for an exact match. Ignored by non-text columns (true by default).

Specifying the :case_sensitive param as FALSE creates other problems, because it uses LOWER(phrase) in the query, which makes the index on the phrase column unusable for the MySQL query optimizer. As we will have thousands and thousands of SAYT Suggestions, doing a full table scan for every new entry is not a scalable option.