Fixed bug where DYM suggestion from bing would get HTML-encoded.
Bing returns spelling suggestions. Their spelling suggestion for 'proctor & gambol' is 'proctor & gamble'. When we encode that and present it to the user, the resulting string is over-encoded, and the search doesn't work as expected.

I don't think Bing is going to return spelling suggestions that contain XSS attacks, but I ran the search page against XSSMe to double check, and it came back fine.