Fixed 4 broken tests that were failing due to DailyQueryStats being empty.
-These tests were failing:
1)
NoMethodError in 'Analytics::HomeController when logged in as an analyst when rendering the monthly reports page should link to the reports on Amazon S3 using SSL if the file exists on S3'
You have a nil object when you didn't expect it!
The error occurred while evaluating nil.strftime
./spec/controllers/analytics/home_controller_spec.rb:107:
./spec/controllers/analytics/home_controller_spec.rb:106:in `each'
./spec/controllers/analytics/home_controller_spec.rb:106:

2)
NoMethodError in 'Analytics::HomeController when logged in as an analyst when rendering the monthly reports page should not link to the English report on S3 if it doesn't exist'
You have a nil object when you didn't expect it!
The error occurred while evaluating nil.strftime
./spec/controllers/analytics/home_controller_spec.rb:117:

3)
NoMethodError in 'Analytics::HomeController when logged in as an analyst when rendering the monthly reports page should not link to the Spanish report on S3 if it doesn't exist'
You have a nil object when you didn't expect it!
The error occurred while evaluating nil.strftime
./spec/controllers/analytics/home_controller_spec.rb:131:

4)
NoMethodError in 'Analytics::HomeController when logged in as an analyst when rendering the monthly reports page should not link to the reports if neither exist'
You have a nil object when you didn't expect it!
The error occurred while evaluating nil.strftime
./spec/controllers/analytics/home_controller_spec.rb:146:
./spec/controllers/analytics/home_controller_spec.rb:145:in `each'
./spec/controllers/analytics/home_controller_spec.rb:145:

- Changed queries schema back to using datetime instead of timestamp for 'timestamp' column