Show top "no results" queries in analytics/queries page
+ created migration for daily_query_noresults_stat model
+ changed query impression logging to log locale&affiliate, and changed format to be more JSON-like
+ fixed Cucumber feature scenario to reflect data on analytics query page, not analytics home page

The intent behind the daily_query_noresults_stat model is similar to the one for Top Movers--- it's a simple data structure that will get populated out-of-band by a daily cron. This code just surfaces it in the analytics area.

The data will be available on a per-affiliate basis, but this commit only shows the data in the usasearch analytics area.

[#4584940]